Cooking System Bugfix & TrashCan System Implemented
Also implemented 1st card set

+++ Assets/[------2 Scripts ------]/1.Manager/sub-Manager/ScoreManager.cs
@@ -12,28 +12,28 @@ public class ScoreManager : MonoBehaviour
         scoreT.text = "Score: 0";
     }
 
-    public void CalculateScore(CardCardGame_Menu menuCard, int ingredientsScore)
+    public void CalculateScore(CardCardGame_Menu menuCard, float ingredientsScore)
     {
         switch(menuCard.menuVariable.mathSymbol)
         {
             case MathSymbol.plus:
-                _score += ingredientsScore + menuCard.menuVariable.amout;
+                _score += (int)(ingredientsScore + menuCard.menuVariable.amout);
             break;
 
             case MathSymbol.minus:
-                _score += ingredientsScore - menuCard.menuVariable.amout;
+                _score += (int)(ingredientsScore - menuCard.menuVariable.amout);
             break;
 
             case MathSymbol.multiply:
-                _score += ingredientsScore * menuCard.menuVariable.amout;
+                _score += (int)(ingredientsScore * menuCard.menuVariable.amout);
             break;
 
             case MathSymbol.divide:
-                _score += ingredientsScore / menuCard.menuVariable.amout;
+                _score += (int)(ingredientsScore / menuCard.menuVariable.amout);
             break;
 
             case MathSymbol.none:
-                _score += ingredientsScore;
+                _score += (int)ingredientsScore;
             break;
         }
